requirements.txt
django==1.8.14
argparse==1.2.1
django-debug-toolbar==1.4
djangorestframework==3.4.5
sqlparse==0.1.18
wheel==0.24.0
wsgiref==0.1.2
django-rest-swagger
psycopg2